// packages/core/src/middleware/material/MaterialConfig.d.ts
export interface MaterialConfig extends SymbolConfig {
    alphaTest: number;
    colorWrite: boolean;
    depthTest: boolean;
    depthWrite: boolean;
    fog: boolean;
    name: string;
    needsUpdate: boolean;
    opacity: number;
    dithering: boolean;
    shadowSide: number | null;
    side: number;
    toneMapped: boolean;
    transparent: boolean;
    visible: boolean;
    blendDst: number;
    blendDstAlpha: number | null;
    blendEquation: number;
    blendEquationAlpha: number | null;
    blending: Blending;
    blendSrc: number;
    blendSrcAlpha: number | null;
    polygonOffset: boolean;
    polygonOffsetFactor: number;
    polygonOffsetUnits: number;
}
// ...
export interface MeshStandardMaterialConfig extends MaterialConfig {
    aoMapIntensity: number;
    bumpScale: number;
    color: string;
    displacementScale: number;
    displacementBias: number;
    emissive: string;
    emissiveIntensity: number;
    envMapIntensity: number;
    flatShading: boolean;
    lightMapIntensity: number;
    metalness: number;
    normalMapType: number;
    refractionRatio: number;
    roughness: number;
    wireframe: boolean;
    wireframeLinecap: string;
    wireframeLinejoin: string;
    roughnessMap: string;
    normalMap: string;
    metalnessMap: string;
    map: string;
    lightMap: string;
    envMap: string;
    emissiveMap: string;
    displacementMap: string;
    bumpMap: string;
    alphaMap: string;
    aoMap: string;
}
// ...
export interface MeshPhysicalMaterialConfig extends MeshStandardMaterialConfig {
    attenuationColor: string;
    attenuationDistance: number;
    clearcoat: number;
    clearcoatNormalScale: Vector2Config;
    clearcoatRoughness: number;
    ior: number;
    reflectivity: number;
    sheen: number;
    sheenRoughness: number;
    sheenColor: string;
    specularIntensity: number;
    specularColor: string;
    thickness: number;
    transmission: number;
    clearcoatMap: string;
    clearcoatNormalMap: string;
    clearcoatRoughnessMap: string;
    sheenRoughnessMap: string;
    sheenColorMap: string;
    specularIntensityMap: string;
    specularColorMap: string;
    thicknessMap: string;
    transmissionMap: string;
}